Add property values to the Jaeger tracing
In Jaeger Tracing span tags Before mediation and after mediation are the same.
If we put a payload factory mediator and change the payload, in span tags
of PayloadFactoryMediator it shows the same payload for beforePayload tag
and the afterPayload tag.
Fix this issue by changing the design of close event handling.

Also Currently, in EI Jager tracing, we can only see the property name.
This commit improves the code to add the property value as well.

Fixes: https://github.com/wso2/product-ei/issues/5218
Fixes: https://github.com/wso2/product-ei/issues/5215