fixed ftl file reference issue
to resolve the issue with loading the ftl file, a flag was added to check whether formatting was done to the template at initialization. if it was not formatted, then the template will be initialized again adding the format. the reason for this is both the templates gets initialized at the startup and only after the proxy or the api call, it will have a format.

public git issue: https://github.com/wso2/api-manager/issues/1520