Fix Passthrough Threads getting stucked due to request message discard
Remove discard request message in ClientWorker since in httpcore-nio already
clears the inPlain buffer. Also set suspendInput state on connection
so that when inputReady state hits it won't go for handleInvalidState.
Also we have decided to proceed with the response regardless of the http status code.
But mark target and source connections to be closed

Fixes: https://github.com/wso2/api-manager/issues/1792