Fix mismatch in axis service and proxy(synapse) service state
For MI we have introduced this feature using a DeploymentInterceptor.
But to enable that we have to add the following service parameter to the proxy

<parameter name="keepServiceHistory">true</parameter>

But there is an issue with MI where it does not properly propagate this state
since be due to a mismatch in axis service and proxy(synapse) service state.

Fixes: https://github.com/wso2/micro-integrator/issues/2981